refactoring : 네이밍 변경
